Attempting to integrate voter data for choropleth map.
Added 2016 voter data for Allegheny County. Attempted to manipulate the precinct data to create a new column in it that would concatenate the municipality name, ward number and precinct number into a single string, which would match up with the field precinct.name on the voter data so it could fit into a choropleth.